Create README.md
A complete research project where we analyzed the malaria dynamics in three connected patches under low migration will be made available on the African Institute for Mathematical Sciences project page. Kindly reach out to me if you find this interesting or wish for us to continue any work of this sort via email.